requirements.txt
# requirements.txt — RAML Platform v5.0 (ADK migration)

# ── Google ADK ────────────────────────────────────────────────────────────────
google-adk>=0.5.0

# ── LiteLLM bridge (ADK uses this for Claude) ────────────────────────────────
litellm>=1.40.0

# ── Anthropic SDK (direct calls in shared/raml_tools.py) ─────────────────────
anthropic>=0.40.0

# ── Vector store & embeddings ─────────────────────────────────────────────────
pinecone-client>=3.0.0
voyageai>=0.2.3

# ── FastAPI server ────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
requests>=2.31.0
